Fix #52. Create MANIFEST.in
Add inside the scripts folder and also the different requirements files.
Update setup.py in order to have a better detection of the package.
Refactor the version in order to have it in deepsphere.__version__
Add a way to solve environment problem in the pyproject.toml (related toproblems that MPecheur had).